+++ web/packages/teleport/src/Discover/Database/DeployService/ManualDeploy/index.ts
@@ -13,5 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import DownloadScript from './DownloadScript';
-export { DownloadScript };
+import ManualDeploy from './ManualDeploy';
+export { ManualDeploy };
